Common test script for v3.1 and vNext
