+++ scripts/schema-test-coverage.sh
@@ -12,10 +12,10 @@ for schemaDir in schemas/v3* ; do
   version=$(basename "$schemaDir")
   echo $version
 
-  node scripts/schema-test-coverage.mjs $schemaDir/schema.yaml tests/$version/pass
+  node scripts/schema-test-coverage.mjs $schemaDir/schema.yaml tests/schemas/$version/pass
 
   echo
 done
 
 echo vNext
-node scripts/schema-test-coverage.mjs src/schemas/validation/schema.yaml tests/vNext/pass
+node scripts/schema-test-coverage.mjs src/schemas/validation/schema.yaml tests/schemas/vNext/pass
