+++ tests/schemas/v3.0/pass/api-with-examples.yaml
