+++ tests/schemas/v3.0/pass/callback-example.yaml
