+++ tests/schemas/v3.0/pass/link-example.yaml
