+++ tests/schemas/v3.0/pass/petstore.yaml
