+++ tests/schemas/v3.0/pass/uspto.yaml
