+++ tests/schemas/v3.1/fail/invalid_schema_types.yaml
