+++ tests/schemas/v3.1/fail/server_enum_empty.yaml
