+++ tests/schemas/v3.1/fail/servers.yaml
