+++ tests/schemas/v3.1/pass/comp_pathitems.yaml
