+++ tests/schemas/v3.1/pass/info_summary.yaml
