+++ tests/schemas/v3.1/pass/mega.yaml
