+++ tests/schemas/v3.1/pass/minimal_hooks.yaml
