+++ tests/schemas/v3.1/pass/minimal_paths.yaml
