+++ tests/schemas/v3.1/pass/non-oauth-scopes.yaml
