+++ tests/schemas/v3.1/pass/path_no_response.yaml
