+++ tests/schemas/v3.1/pass/path_var_empty_pathitem.yaml
