+++ tests/schemas/v3.1/pass/schema.yaml
