+++ tests/schemas/v3.1/pass/servers.yaml
