+++ tests/schemas/v3.1/pass/valid_schema_types.yaml
