+++ tests/schemas/vNext/fail/invalid_schema_types.yaml
