+++ tests/schemas/vNext/fail/no_containers.yaml
