+++ tests/schemas/vNext/fail/server_enum_empty.yaml
