+++ tests/schemas/vNext/fail/servers.yaml
