+++ tests/schemas/vNext/pass/comp_pathitems.yaml
