+++ tests/schemas/vNext/pass/info_summary.yaml
