+++ tests/schemas/vNext/pass/mega.yaml
