+++ tests/schemas/vNext/pass/minimal_comp.yaml
