+++ tests/schemas/vNext/pass/minimal_hooks.yaml
