+++ tests/schemas/vNext/pass/minimal_paths.yaml
