+++ tests/schemas/vNext/pass/non-oauth-scopes.yaml
