+++ tests/schemas/vNext/pass/path_no_response.yaml
