+++ tests/schemas/vNext/pass/path_var_empty_pathitem.yaml
