+++ tests/schemas/vNext/pass/schema.yaml
