+++ tests/schemas/vNext/pass/servers.yaml
