+++ tests/schemas/vNext/pass/valid_schema_types.yaml
